chore(sanity): clean up version context
